% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCost.R
\name{computeCost}
\alias{computeCost}
\title{Single Cell compute cost Tab Server}
\usage{
computeCost(
  input,
  output,
  session,
  analysis_type,
  combinedConsMarkersTable,
  combinedMarkersTable,
  seuratObjectsList,
  costPerMil,
  depthPerCell
)
}
\arguments{
\item{analysis_type}{Reactive value containing type of analysis (integrated vs single analysis)}

\item{combinedConsMarkersTable}{Reactive value containing a table that lists conserved markers per cluster per
subset if the "find conserved markers" option was selected}

\item{seuratObjectsList}{Reactive value containing list of downsampled seurat objects
with reduced dimensions (PCA data), scaled counts, and cluster labels that corresponds across subsets}

\item{costPerMil}{Reactive value containing cost of sequencing one million reads}

\item{depthPerCell}{Reactive value containing the depth of sequencing per cell selected by the user}
}
\description{
Single Cell compute cost Tab Server
}
