% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConservedMarkers.R
\name{conservedMarkersUI}
\alias{conservedMarkersUI}
\title{Single Cell conserved markers Tab UI}
\usage{
conservedMarkersUI(id)
}
\value{
None
}
\description{
Single Cell conserved markers Tab UI
}
