% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConservedMarkers.R
\name{findConservedMarkers}
\alias{findConservedMarkers}
\title{Single Cell conserved markers Server}
\usage{
findConservedMarkers(
  input,
  output,
  session,
  seuratObjectsList,
  selectedNumGenes
)
}
\arguments{
\item{seuratObjectsList}{Reactive value containing list of downsampled seurat objects}

\item{selectedNumGenes}{Reactive value containing user input for number of top genes to test for per cluster per subset}
}
\value{
Returns a reactive value containing a table that lists conserved markers per cluster per subset
}
\description{
Single Cell conserved markers Server
}
