% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEgenes.R
\name{findDEgenes}
\alias{findDEgenes}
\title{Single Cell differential expression Server}
\usage{
findDEgenes(
  input,
  output,
  session,
  seuratObjectsList,
  dataset1_name,
  dataset2_name,
  selectedNumGenes
)
}
\arguments{
\item{seuratObjectsList}{Reactive value containing list of downsampled seurat objects}

\item{dataset1_name}{Reactive value holding project name for second dataset}

\item{selectedNumGenes}{Reactive value containing user input for number of top genes to test for per cluster per subset}
}
\value{
Returns a Reactive value containing a table that lists deferentially expressed genes present in each subset
}
\description{
Single Cell differential expression Server
}
