% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerGenes.R
\name{findMarkerGenes}
\alias{findMarkerGenes}
\title{Single Cell marker genes Server}
\usage{
findMarkerGenes(input, output, session, seuratObjectsList, selectedNumGenes)
}
\arguments{
\item{seuratObjectsList}{Reactive value containing list of downsampled seurat objects}

\item{selectedNumGenes}{Reactive value containing user input for number of top genes to test for per cluster per subset}
}
\value{
Returns a Reactive value containing a table listing marker genes present in each subset
}
\description{
Single Cell marker genes Server
}
