% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{read_10x}
\alias{read_10x}
\title{#' Single Cell 10x data reading Tab Server}
\usage{
read_10x(input, output, session, dataset_name, uploadsDF)
}
\arguments{
\item{dataset_name}{Reactive value containing dataset name}

\item{uploadsDF}{Reactive value containing dataframe having name, size, type and datapath of the uploaded files}
}
\value{
Returns a Reactive value containing seurat object
}
\description{
#' Single Cell 10x data reading Tab Server
}
