% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{read_h5}
\alias{read_h5}
\title{#' Single Cell .h5 data reading Tab Server}
\usage{
read_h5(input, output, session, dataset_name, path)
}
\arguments{
\item{dataset_name}{Reactive value containing dataset name}

\item{path}{Reactive value containing path to matrix files}
}
\value{
Returns a Reactive value containing seurat object
}
\description{
#' Single Cell .h5 data reading Tab Server
}
