% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClustering.R
\name{renameClusters}
\alias{renameClusters}
\title{Function to reassign cluster labels across downsampled subsets}
\usage{
renameClusters(seuratObjectsList)
}
\arguments{
\item{seuratObjectsList}{Reactive value containg list of downsampled seurat objects
with reduced dimensions (PCA data) and scaled counts}
}
\value{
Returns a Reactive value containing list of downsampled seurat objects
 with reassigned cluster labels that correspoond across downsampled subsets
}
\description{
Function to reassign cluster labels across downsampled subsets
}
