% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClustering.R
\name{scClustering}
\alias{scClustering}
\title{Single Cell Cluster Tab Server}
\usage{
scClustering(input, output, session, seurat, mito, res, dataset1_name)
}
\arguments{
\item{seurat}{Reactive value containing seurat object}

\item{mito}{Reactive value containing user input for mitochondrial genes pattern}

\item{res}{Reactive value containing user input for clustering resolution}

\item{dataset1_name}{Reactive value containing user input of uploaded dataset name}
}
\value{
Returns a Reactive value containing list of downsampled seurat objects
 with reduced dimensions (PCA data) and scaled counts
}
\description{
Single Cell Cluster Tab Server
}
