% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{scIntegrate}
\alias{scIntegrate}
\title{#' Single Cell Integration Tab Server}
\usage{
scIntegrate(input, output, session, seurat, seurat2, mito, res)
}
\arguments{
\item{seurat}{Reactive value containing seurat object}

\item{seurat2}{Reactive value containing seurat object}

\item{mito}{Reactive value containing user input for mitochondrial genes pattern}

\item{res}{Reactive value containing user input for clustering resolution}

\item{dataset}{Reactive value containg user input of uploaded dataset name}
}
\value{
Returns a Reactive value containing list of downsampled seurat objects
 with reduced dimensions (PCA data) and scaled counts
}
\description{
#' Single Cell Integration Tab Server
}
