% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClustering.R
\name{scVarGenes}
\alias{scVarGenes}
\title{Single Cell Variable genes Tab Server}
\usage{
scVarGenes(input, output, session, seuratObjectsList)
}
\arguments{
\item{seuratObjectsList}{Reactive value containg list of downsampled seurat objects
with reduced dimensions (PCA data) and reduced dimensions}
}
\description{
Single Cell Variable genes Tab Server
}
