% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConservedMarkers.R
\name{tableParserCMG}
\alias{tableParserCMG}
\title{Function that iterates through cell markers and record their presence/absence per subset.}
\usage{
tableParserCMG(conservedMarkers_Tables_List)
}
\arguments{
\item{conservedMarkers_Tables_List}{Reactive value containing a list of tables that lists the conserved markers present in each subset}
}
\value{
Returns a Reactive value listing conserved marker genes resolved per subset
}
\description{
Function that iterates through cell markers and record their presence/absence per subset.
}
