% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEgenes.R
\name{tableParserDEG}
\alias{tableParserDEG}
\title{Function that iterates through differentially expressed genes and records their presence/absence per cluster per subset.}
\usage{
tableParserDEG(DE_Tables_List)
}
\arguments{
\item{DE_Tables_List}{Reactive value containing a list of tables that list the deferentially expressed genes
present per cluster per subset}
}
\value{
Returns a Reactive value listing conserved marker genes resolved per cluster per subset
}
\description{
Function that iterates through differentially expressed genes and records their presence/absence per cluster per subset.
}
