% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkerGenes.R
\name{tableParserMG}
\alias{tableParserMG}
\title{Function that iterates through cell markers and record their presence/absence per subset.}
\usage{
tableParserMG(TablesList)
}
\arguments{
\item{TablesList}{Reactive value containing a list of tables that list the marker genes present prer cluster per
subset}
}
\value{
Returns a Reactive value listing marker genes resolved per cluster per subset
}
\description{
Function that iterates through cell markers and record their presence/absence per subset.
}
